Added warning on Progress Bubble
In progress bubble, when player mass gets too close to level down size,
the circle will flash red. There are option to change the color it
flashing, the transition animation, and the threshold at which it starts
to warn the player. Level down code should be put in here, so we have a
centralized place to do the leveling.