Added Final Score Scene
The scene FinalScore, which displays after game over is created. It
takes the score and color from the main game, as such, Score.Instance
and PlayerCharacter.Instance are not destroyed until the set up for the
FinalScore scene has been completed. The Director gameObject, which
handles the order in which events occur, may need linking up. It should
have public variables that take in gameObjects within the scene as
labeled. Currently, the mass of the player is inconsequential to the
way that the FinalScore scene's particle effects are rendered. The
FinalScore scene CANNOT be run by itself. To access it, go to the main
game scene and press 'q'.